Extract folderItemSpec to a new file
